Remove pointless ToggleButtonApp class.
This class is not really needed, it's just a container. And is
only used once.  So move it there.  Note you can nest layouts,
no need to have a separate QWidget to do it.

It's also weird to connect to children like that. Instead of
having the buttons in a children widget, it is the view and
control to whether show cross that should be the children.
(todo next, move that stuff into the CameraView class)