Remove duplicated help text and input checking (issue #9)
The command line arguments are already parsed by argparse.  In
case of errors, argparse will print help text.  So, remove the
manual check for command line arguments in __main__.

The setuptools entry points do not call the named function with
the command line arguments, so we still need a wrapper function
that does this (__main__).  However, we still want a Python
function that can be called with a list of arguments mimicking
the command line usage (main).