README: markdown style, usage fixes, and drop not essential info.
Use markdown style for sectioning, code blocks, and argument names
(backticks, not quotes).  Also, the program is named BeamDelta, not
BeamDelta, at least one argument is non optional so no [].  Drop
mention of exposure which is not essential, link to --help instead.
Copyright holders and year is not something for a README which should
have the most important stuff to get started.  Remove python version
and packages requirements which are duplicated from setup.py and
already out of date.